Switch from numpy aliases to builtin types

+++ scvelo/core/tests/test_models.py
@@ -21,7 +21,7 @@ class TestSplicingDynamics:
             st.floats(min_value=0, allow_infinity=False), min_size=2, max_size=2
         ),
         t=arrays(
-            np.float,
+            float,
             shape=st.integers(min_value=1, max_value=100),
             elements=st.floats(
                 min_value=0, max_value=1e3, allow_infinity=False, allow_nan=False
